requirements.txt
anthropic>=0.45.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
pydantic>=2.0.0
